# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Traffic safety community portal backend.",
        "title": "Obvio Community Portal API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/annotations/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "annotations"
                ],
                "summary": "Update an annotation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Annotation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Annotation data",
                        "name": "annotation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/annotation.Annotation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/annotation.Annotation"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "annotations"
                ],
                "summary": "Delete an annotation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Annotation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/events/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get event by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/event.Event"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Update an event",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Event data",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/event.Event"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/event.Event"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "events"
                ],
                "summary": "Delete an event",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/events/{id}/annotations": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "annotations"
                ],
                "summary": "List annotations for an event",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/annotation.Annotation"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "annotations"
                ],
                "summary": "Create an annotation for an event",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Annotation data",
                        "name": "annotation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/annotation.Annotation"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/annotation.Annotation"
                        }
                    }
                }
            }
        },
        "/events/{id}/upload": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Upload video for an event",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Video file",
                        "name": "video",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/portals/{id}/insights": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "insights"
                ],
                "summary": "Get insights for a portal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Portal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usecase.InsightsResult"
                        }
                    }
                }
            }
        },
...
        "/trials/{id}/events": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "List events for a trial",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Trial ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/event.Event"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Create an event for a trial",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Trial ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Event data",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/event.Event"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/event.Event"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "annotation.Annotation": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "event_id": {
                    "type": "integer"
                },
                "frame_time": {
                    "type": "number"
                },
                "height": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "width": {
                    "type": "number"
                },
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        },
        "event.Event": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "duration": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "is_violation": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                },
                "trial_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "video_url": {
                    "type": "string"
                },
                "violation_type": {
                    "type": "string"
                }
            }
        },
...
        "usecase.EventsByDayItem": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                }
            }
        },
        "usecase.EventsByHourItem": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "hour": {
                    "type": "integer"
                }
            }
        },
        "usecase.InsightsResult": {
            "type": "object",
            "properties": {
                "events_by_day": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.EventsByDayItem"
                    }
                },
                "events_by_hour": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.EventsByHourItem"
                    }
                },
                "peak_day": {
                    "type": "string"
                },
                "peak_hour": {
                    "type": "integer"
                },
                "total_events": {
                    "type": "integer"
                },
                "total_violations": {
                    "type": "integer"
                },
                "violation_breakdown": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.ViolationBreakdownItem"
                    }
                },
                "violation_rate": {
                    "type": "number"
                }
            }
        },
        "usecase.ViolationBreakdownItem": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    }
}